+++ src/crypto.ts
@@ -1,12 +1,12 @@
 // Isomorphic pattern to load Web Crypto in both browsers and Node.js
-async function loadCrypto() {
+async function loadCrypto(): Promise<Crypto> {
     if (globalThis && globalThis.crypto) {
         // Running in browsers from 2018 and newer.
         return new Promise((resolve, _reject) => resolve(globalThis.crypto))
     } else {
         // Running in Node.js >= 15
         const cryptoLocal = await import('crypto')
-        return cryptoLocal.webcrypto
+        return ((cryptoLocal.webcrypto as unknown) as Crypto)
     }
 }